Move all alloc integration tests to a new alloctests crate
