+++ library/alloctests/benches/slice.rs
