+++ library/alloctests/benches/str.rs
