+++ library/alloctests/lib.rs
@@ -1,0 +1,1 @@
+// Intentionally left empty.
