+++ library/alloctests/tests/autotraits.rs
