+++ library/alloctests/tests/const_fns.rs
