+++ library/alloctests/tests/thin_box.rs
