+++ library/alloctests/tests/vec_deque.rs
